A fast reading mode is added for MINST data
I added one mode to save extracted files and load them again in order to reduce the extraction time, which is now required for all the time. There is trade-off between the size of a saved file and the speed of loading. However, the fast loading mode will give some benefit for not large data.